requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
benepar==0.1.2
bert-extractive-summarizer==0.9.0
blis==0.7.6
boto3==1.21.32
botocore==1.24.32
cachetools==5.0.0
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cymem==2.0.6
Cython==0.29.28
Django==3.2.13
django-embed-video==1.4.2
djangorestframework==3.13.1
# en-core-web-sm==3.0.0
filelock==3.6.0
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.44.0
h11==0.13.0
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.11.3
install==1.3.5
Jinja2==3.0.3
jmespath==1.0.0
joblib==1.1.0
langcodes==3.3.0
Markdown==3.3.6
MarkupSafe==2.1.1
multidict==6.0.2
murmurhash==1.0.6
nltk==3.7
numpy==1.21.6
oauthlib==3.2.0
outcome==1.1.0
packaging==21.3
pandas==1.1.5
pathy==0.6.1
Pillow==9.1.0
preshed==3.0.6
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pyDeprecate==0.3.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytz==2021.3
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.7.3
selenium==4.1.3
sentence-transformers==2.2.0
sentencepiece==0.1.96
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
sqlparse==0.4.2
srsly==2.4.2
summa==1.2.0
summarizer==0.0.7
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thinc==8.0.15
threadpoolctl==3.1.0
tokenizers==0.11.6
torch==1.11.0
torchmetrics==0.7.3
torchvision==0.12.0
tqdm==4.63.0
transformers==4.16.2
trio==0.20.0
trio-websocket==0.9.2
typer==0.4.0
typing-extensions==4.1.1
urllib3==1.26.8
wasabi==0.9.0
Werkzeug==2.1.0
wsproto==1.1.0
yarl==1.7.2
youtube-transcript-api==0.4.4
zipp==3.7.0
ipywidgets==7.7.0
